refactor(routes): move session domain into routes/session/ subpackage
Slice 2q of the route-domain reorganization (#4082/#4071). Moves
session_routes.py (1346 lines) into routes/session/, leaving a
backward-compat sys.modules shim. Pure file reorganization, no behavior
change.

The shim uses sys.modules replacement so the module-level `router`
replacement pattern (test_history_compact_tool_calls.py,
test_history_display_model_hydration.py), setattr monkeypatching, the
`from routes import session_routes` parent-attribute form, and
inspect.getsource all reach the canonical module.

Tests updated to also clear the canonical module path when clearing the
legacy path (same pattern as #5898): the sys.modules shim caches the
canonical module, so clearing only "routes.session_routes" doesn't force
re-execution — the shim hits the cached canonical and stubbed imports
never rebind. Affected: test_session_ghost_delete.py,
test_blind_compare_redaction.py, test_session_owner_attribution.py,
test_security_regressions.py (_import_session_routes_for_filename).

Two source-introspection test sites repointed:
- test_aux_llm_owner_scope.py
- test_security_regressions.py (session html export test)

Adds tests/test_session_routes_shim.py to pin the sys.modules shim
contract.

Verified: compileall clean; full suite 5791 passed, 3 skipped.
