+++ routes/session/__init__.py
@@ -1,0 +1,5 @@
+"""Session route domain package (slice 2q, #4082/#4071).
+
+Contains session_routes.py, migrated from the flat routes/ directory.
+Backward-compat shim at routes/session_routes.py re-exports from here.
+"""
